refactor: getDatasetAtEvent, getElementAtEvent and getElementsAtEvent props are removed
getDatasetAtEvent, getElementAtEvent and getElementsAtEvent props are removed, utils with the same
names can used instead

BREAKING CHANGE: getDatasetAtEvent, getElementAtEvent and getElementsAtEvent props are removed,
utils with the same names can used instead
