+++ src/index.tsx
@@ -1,3 +1,8 @@
 export type { ChartProps } from './types';
 export * from './chart';
 export * from './typedCharts';
+export {
+  getDatasetAtEvent,
+  getElementAtEvent,
+  getElementsAtEvent,
+} from './utils';
