+++ stories/Chart.tsx
@@ -1,8 +1,19 @@
-import React, { useState, useEffect, useReducer } from 'react';
+import React, {
+  MouseEvent,
+  useRef,
+  useState,
+  useEffect,
+  useReducer,
+} from 'react';
 import 'chart.js/auto';
-import { InteractionItem } from 'chart.js';
+import type { Chart as ChartJS, InteractionItem } from 'chart.js';
 import 'chartjs-adapter-date-fns';
-import { Chart } from '../src';
+import {
+  Chart,
+  getDatasetAtEvent,
+  getElementAtEvent,
+  getElementsAtEvent,
+} from '../src';
 import * as multitypeChart from '../sandboxes/chart/multitype/App';
 import * as eventsChart from '../sandboxes/chart/events/App';
 import * as data from './Chart.data';
@@ -52,7 +63,7 @@ export const ClickEvents = ({
   data,
   ...args
 }) => {
-  const getDatasetAtEvent = (dataset: InteractionItem[]) => {
+  const datasetAtEvent = (dataset: InteractionItem[]) => {
     if (!dataset.length) return;
 
     const datasetIndex = dataset[0].datasetIndex;
@@ -60,7 +71,7 @@ export const ClickEvents = ({
     onDatasetClick(data.datasets[datasetIndex].label);
   };
 
-  const getElementAtEvent = (element: InteractionItem[]) => {
+  const elementAtEvent = (element: InteractionItem[]) => {
     if (!element.length) return;
 
     const { datasetIndex, index } = element[0];
@@ -68,21 +79,34 @@ export const ClickEvents = ({
     onElementClick(data.labels[index], data.datasets[datasetIndex].data[index]);
   };
 
-  const getElementsAtEvent = (elements: InteractionItem[]) => {
+  const elementsAtEvent = (elements: InteractionItem[]) => {
     if (!elements.length) return;
 
     onElementsClick(elements);
   };
 
+  const chartRef = useRef<ChartJS>(null);
+
+  const onClick = (event: MouseEvent<HTMLCanvasElement>) => {
+    const { current: chart } = chartRef;
+
+    if (!chart) {
+      return;
+    }
+
+    datasetAtEvent(getDatasetAtEvent(chart, event));
+    elementAtEvent(getElementAtEvent(chart, event));
+    elementsAtEvent(getElementsAtEvent(chart, event));
+  };
+
   return (
     <Chart
       {...args}
+      ref={chartRef}
       type='bar'
       options={options}
       data={data}
-      getDatasetAtEvent={getDatasetAtEvent}
-      getElementAtEvent={getElementAtEvent}
-      getElementsAtEvent={getElementsAtEvent}
+      onClick={onClick}
     />
   );
 };
